# Backend/docs/openapi.yaml
openapi: 3.0.3

info:
  title: File Transfer Service API
  description: |
    A RESTful API for uploading, downloading, listing, and deleting files.

    File metadata is stored in PostgreSQL while file contents are stored in MinIO object storage.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local Development Server

tags:
  - name: Files
    description: File management operations

paths:

  /files:

    post:
      tags:
        - Files
      summary: Upload a file
      description: Upload a file using multipart/form-data.
      operationId: uploadFile

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary

      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"

        "400":
          description: Invalid request

        "500":
          description: Internal server error

    get:
      tags:
        - Files
      summary: List uploaded files
      description: Returns metadata for all uploaded files.
      operationId: listFiles

      responses:
        "200":
          description: List of uploaded files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"

        "500":
          description: Internal server error

  /files/{id}:

    get:
      tags:
        - Files
      summary: Download a file
      description: Downloads the file associated with the provided ID.
      operationId: downloadFile

      parameters:
        - name: id
          in: path
          required: true
          description: Unique file identifier
          schema:
            type: string

      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "404":
          description: File not found

        "500":
          description: Internal server error

    delete:
      tags:
        - Files
      summary: Delete a file
      description: Deletes both the file metadata and the object stored in MinIO.
      operationId: deleteFile

      parameters:
        - name: id
          in: path
          required: true
          description: Unique file identifier
          schema:
            type: string

      responses:
        "204":
          description: File deleted successfully

        "404":
          description: File not found

        "500":
          description: Internal server error

components:

  schemas:

    File:
      type: object

      properties:

        id:
          type: string
          example: "0c75df55-3c86-4978-a57d-d93f4d5eaef4"

        fileName:
          type: string
          example: "resume.pdf"

        objectKey:
          type: string
          example: "uploads/0c75df55-3c86-4978-a57d-d93f4d5eaef4-resume.pdf"

        size:
          type: integer
          format: int64
          example: 245678

        contentType:
          type: string
          example: "application/pdf"

        status:
          type: string
          example: "Uploaded"

        uploadedAt:
          type: string
          format: date-time
          example: "2026-07-28T11:20:43Z"